requirements.txt
python-dotenv~=1.0.1
folium~=0.18.0
geojson~=3.1.0
numpy~=2.1.3
openrouteservice~=2.3.3
shapely~=2.0.6